% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_counts.R
\name{calculate_counts}
\alias{calculate_counts}
\title{Calculate ActiGraph activity counts from raw acceleration data}
\usage{
calculate_counts(raw, epoch, lfe_select = FALSE, tz = "UTC", verbose = FALSE)
}
\arguments{
\item{raw}{data frame of raw acceleration data obtained from
\code{\link[read.gt3x]{read.gt3x}}}

\item{epoch}{The epoch length for which the counts should be summed.}

\item{lfe_select}{Apply the Actigraph Low Frequency Extension filter, Default: FALSE}

\item{tz}{the desired timezone (defaults to \code{UTC})}

\item{verbose}{Print the progress of the Actigraph raw data conversion to counts, Default: FALSE.}
}
\description{
Calculate ActiGraph activity counts from raw acceleration data
}
\examples{
f <- system.file("extdata/example.gt3x", package = "agcounts")
d <- read.gt3x::read.gt3x(f, asDataFrame = TRUE, imputeZeroes = TRUE)
calculate_counts(d, 60)
}
