% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcounts.R
\name{get_counts}
\alias{get_counts}
\title{get_counts}
\usage{
get_counts(
  path,
  frequency,
  epoch,
  lfe_select = FALSE,
  write.file = TRUE,
  return.data = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path}{Full path name to the GT3X File}

\item{frequency}{Sampling frequency in hertz. Sampling frequency must be a multiple of 10 between 30 and 100 hertz.}

\item{epoch}{The epoch length for which the counts should be summed.}

\item{lfe_select}{Apply the Actigraph Low Frequency Extension filter, Default: FALSE}

\item{write.file}{Export a CSV file of the counts, Default: TRUE}

\item{return.data}{Return the data frame to the R Global Environment, Default: FALSE}

\item{verbose}{Print the progress of the Actigraph raw data conversion to counts, Default: FALSE.}
}
\value{
Returns a CSV file if write.file is TRUE or a data frame if return.data is TRUE
}
\description{
Main function to extract counts from the Actigraph GT3X Files.
}
\details{
Main function to extract counts from the Actigraph GT3X Files.
}
\examples{
agcounts::get_counts(path = "filename.gt3x", frequency = 60, epoch = 60, lfe_select = FALSE, write.file = FALSE, return.data = TRUE)
}
\seealso{
\code{\link[read.gt3x]{read.gt3x}},\code{\link[read.gt3x]{read_gt3x}}
}
