% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{upSampleC}
\alias{upSampleC}
\title{Upsample data if frequency is 30, 60, or 90 Hertz}
\arguments{
\item{X}{NumericMatrix. The matrix that will be upsampled.}

\item{b_fp}{double. A factor to be multiplied by in the upsampling process.}
}
\description{
Upsample data if frequency is 30, 60, or 90 Hertz
}
\keyword{internal}
