% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agread.R
\name{agread}
\alias{agread}
\title{Read in raw acceleration data}
\usage{
agread(
  raw,
  verbose = FALSE,
  parser = c("pygt3x", "ggir", "gt3x"),
  tz = "UTC",
  ...
)
}
\arguments{
\item{raw}{Path name to the GT3X file or the dataset with columns time, X, Y, and Z axis}

\item{verbose}{Print the read method, Default: FALSE.}

\item{parser}{The parser to use when reading in the data. Parser values include pygt3x, ggir, and gt3x (uncalibrated reader)}

\item{tz}{the desired timezone, Default: \code{UTC}}

\item{...}{Additional arguments to pass into the agread S3 generic function}
}
\value{
Returns the raw acceleration signal
}
\description{
This generic function reads in calibrated raw acceleration data
    from the python module pygt3x or the R GGIR package or uncalibrated data
    from the read.gt3x package.
}
\details{
This generic function reads in calibrated raw acceleration data
    from the python module pygt3x or the R read.gt3x and GGIR packages.
}
\examples{
\dontrun{
   agread(system.file("extdata/example.gt3x", package = agcounts))
}
}
