% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agread.R
\name{agread}
\alias{agread}
\title{Read in raw acceleration data}
\usage{
agread(
  path,
  parser = c("pygt3x", "ggir", "uncalibrated"),
  tz = "UTC",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{path}{Path name to the GT3X file or the dataset with columns time, X, Y, and Z axis}

\item{parser}{The parser to use when reading in the data. Parser values include pygt3x, ggir, and uncalibrated readers}

\item{tz}{the desired timezone, Default: \code{UTC}}

\item{verbose}{Print the read method, Default: FALSE.}

\item{...}{Additional arguments to pass into the agread function}
}
\value{
Returns the calibrated or uncalibrated raw acceleration data
}
\description{
This generic function reads in calibrated raw acceleration data
    from the python module pygt3x or the R GGIR package or uncalibrated data
    from the read.gt3x package.
}
\details{
This function reads in calibrated raw acceleration data from GGIR or pygt3x.
}
\examples{
\dontrun{
   agread(system.file("extdata/example.gt3x", package = "agcounts"), parser = "pygt3x")
}
}
\seealso{
\code{\link[GGIR]{g.calibrate}}
 \code{\link[read.gt3x]{read.gt3x}}
}
