% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{get_counts}
\usage{
get_counts(
  path,
  epoch,
  lfe_select = FALSE,
  write.file = FALSE,
  return.data = TRUE,
  verbose = FALSE,
  tz = "UTC",
  ...
)
}
\arguments{
\item{path}{Full path name to the GT3X File}

\item{epoch}{The epoch length for which the counts should be summed.}

\item{lfe_select}{Apply the Actigraph Low Frequency Extension filter, Default: FALSE}

\item{write.file}{Export a CSV file of the counts, Default: FALSE}

\item{return.data}{Return the data frame to the R Global Environment, Default: TRUE}

\item{verbose}{Print the progress of the Actigraph raw data conversion to counts, Default: FALSE.}

\item{tz}{the desired timezone, Default: \code{UTC}}

\item{...}{arguments passed to \code{\link[data.table]{fwrite}}}
}
\value{
Returns a CSV file if write.file is TRUE or a data frame if return.data is TRUE
}
\description{
Main function to extract counts from the Actigraph GT3X Files.
}
\details{
Main function to extract counts from the Actigraph GT3X Files.
}
\examples{
if(interactive()){
 get_counts(
  path = system.file("extdata/example.gt3x", package = "agcounts"),
  epoch = 60, lfe_select = FALSE,
  write.file = FALSE, return.data = TRUE
  )
  }
}
\seealso{
\code{\link[read.gt3x]{read.gt3x}}
}
